reorganize game files
