+++ react-frontend/src/Game/GameTypes.ts
